feat: generate a initial csv file to the project
This is generated via a python script, and will initially be set to 300
papers that will be queried. From the generated XML file, a notebook
block then makes it into a CSV file that can easily be read with Neo4J.
However, there might be few references in this graph database, so later
there likely must be made additional inserts.